Reorganized package structure to allow for new components
