+++ src/tests/App.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from '../app/App';
 
 test('renders learn react link', () => {
   render(<App />);
